+++ src/boot/ksdk1.1.0/devCCS811.c
@@ -168,7 +168,7 @@ writeSensorRegisterCCS811(uint8_t deviceRegister, uint8_t *payload, uint16_t men
 }
 
 WarpStatus
-configureSensorCCS811(uint8_t *payloadMEAS_MODE, uint8_t menuI2cPullupValue)
+configureSensorCCS811(uint8_t *payloadMEAS_MODE, uint16_t menuI2cPullupValue)
 {
 	WarpStatus	status1, status2;
